refactor: Added a new middle level IR (MIR) in between the AST and the backend.
Datatypes are converted to primitive datatypes such as iN, uN, fN and such.
The middle-end converts the AST to the MIR. The backend only works with the MIR. This reduces the frontend details bleeding into the backend and reduces coupling between the frontend and backend.

+++ tests/codeGen/test_var.c
@@ -1,9 +1,9 @@
 int main(){
-    long int a = 0, b;
+    int a = 0, b;
     a = 12;
     b = 12/3;
     
-    long int *ap;
+    int *ap;
     ap = &a;
 
     *ap = 20;
